% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirvash-package.R
\docType{package}
\name{nirvash-package}
\alias{nirvash}
\alias{nirvash-package}
\title{nirvash: Put Texts 'writing-mode: tb-rl';}
\description{
'tb-rl' styled text block component
    powered by nehan.js (v5.5.10).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/nirvash}
  \item Report bugs at \url{https://github.com/paithiov909/nirvash/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
