% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirvash-package.R, R/nirvash.R
\docType{package}
\name{nirvash}
\alias{nirvash}
\alias{nirvash-package}
\title{nirvash: Vertical-Styled Text Block Component}
\usage{
nirvash(
  context,
  split = "\\n",
  serif = TRUE,
  writing_mode = c("tbrl", "lrtb"),
  custom_style = "",
  width = "100\%",
  height = "90vh"
)
}
\arguments{
\item{context}{text to disply on widget.}

\item{split}{stirng with that split the context and wrap as a paragraph.}

\item{serif}{boolean. if true, set an additional style on `.serif` class elements.}

\item{writing_mode}{choose from "tbrl" or "lrtb".}

\item{custom_style}{user's own defined sytle that will be attached onto component's wrapper directly.}
}
\value{
object returned from htmlwidgets::createWidget()
}
\description{
'vertical-rl' styled text block component
    powered by nehan.js (v5.5.10).

`vertical-rl` styled component
powered by nehan.js (v5.5.10).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/nirvash}
  \item Report bugs at \url{https://github.com/paithiov909/nirvash/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
