% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirvash.R
\name{nirvash}
\alias{nirvash}
\title{Nirvash HTML Widget}
\usage{
nirvash(
  context,
  split = "\\n",
  serif = TRUE,
  writing_mode = c("tbrl", "lrtb"),
  custom_style = "",
  width = "100\%",
  height = "90\%"
)
}
\arguments{
\item{context}{text to disply on widget.}

\item{split}{string with that nirbash splits the context into paragraphs.}

\item{serif}{boolean. if true, set an additional style onto `.serif` class elements.}

\item{writing_mode}{choose from "tbrl" or "lrtb".}

\item{custom_style}{user's own defined sytle that will be attached onto component's wrapper directly.}
}
\value{
object returned from htmlwidgets::createWidget()
}
\description{
`vertical-rl` styled component
powered by nehan.js (v5.5.10).
}
