% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirvash.R
\name{nirvash}
\alias{nirvash}
\title{Nirvash HTML Widget}
\usage{
nirvash(
  context,
  split = "\\n",
  serif = TRUE,
  writing_mode = c("tbrl", "lrtb"),
  custom_style = "",
  width = "100\%",
  height = "90\%"
)
}
\arguments{
\item{context}{Text to display on widget.}

\item{split}{String with that nirvash splits the context into paragraphs.}

\item{serif}{Boolean. If true, sets an additional style onto `.serif` class elements.}

\item{writing_mode}{Choose one of "tbrl" or "lrtb".}

\item{custom_style}{User's own defined style that will be attached onto component's wrapper.}

\item{width}{Width of components.}

\item{height}{Height of components.}
}
\value{
Object returned from \code{htmlwidgets::createWidget()}.
}
\description{
`tb-rl` styled component
powered by nehan.js (v5.5.10).
The widgets do not support resizing.
}
